move colors configurations to  hepler file
