+++ react/features/riff-dashboard-page/src/components/Dashboard/Metrics/StackedBarGraph.js
@@ -20,11 +20,8 @@ import * as am4charts from '@amcharts/amcharts4/charts';
 import * as d3 from 'd3-array';
 
 import { ScaleLoader } from 'react-spinners';
-
-import {
-    Colors,
-    logger,
-} from 'libs/utils';
+import { logger } from 'libs/utils';
+import { Colors } from './colorsHelpers';
 
 import { EStatus, GraphConfigs, GraphTypes } from 'libs/utils/constants';
